MBKeyboardKey: don't do full keyboard redraw on release when in Caps mode
If a key has a modifier then on release we have to redraw the whole keyboard
because on release the modifier is reset, so the kbd layout changes. The
Caps modifier is an exception, since it stays in place until it is explicitely
cancelled, but we were still scheduling a full redraw; if the user drags a
finger across the kbd and the HW is not fast enough to keep up with the
redraws, this becomes rather obvious.