MbGtkKeyboard: ensure correct backbuffer setup for map->unmap->map sequence
When the widget is going from an unmapped to a mapped state, if it was
previously mapped at the same size, it is possible for the backend to end up
attached to a stale drawable. So when the widget gets unmapped, set a flag
to force fresh backbuffer set up on any future map at an identical size.

(If the new map happens at a different size, this triggers complete relayout,
which included backbuffer set up.)

Signed-off-by: Tomas Frydrych <tomas@sleepfive.com>