MBKeyboardUI: ensure correct backing pixmap on map events
When the kbd window maps, it is possible that it is not using the current
backing pixmap. The preceeding commit has was attempting to address this
problem by forcing backbuffer setup after an unmap event, but this solution
fails when the backbuffer pixmap is not attached on the initial map, which
seems to happen with more recent versions of gtk.

Signed-off-by: Tomas Frydrych <tomas@tomtain.com>